* Mise en place de threads pour l'enregistrement des campagnes * Mise en francais

+++ src/main/java/net/courtanet/arato/tsunami/dao/Dao.java
@@ -4,17 +4,22 @@ import net.courtanet.arato.tsunami.cluster.CassandraCluster;
 
 public abstract class Dao {
 
-	public void createTable() {
-		System.out.println("Suppression ancienne table " + getTable());
-		CassandraCluster.getInstance().getSession()
-				.execute("DROP TABLE IF EXISTS " + getTable() + ";");
+	public void creerTable() {
+		if (doitEtreSupprimee()) {
+			System.out.println("Suppression ancienne table " + getTable());
+			CassandraCluster.getInstance().getSession()
+					.execute("DROP TABLE IF EXISTS " + getTable() + ";");
+		}
 		System.out.println("Création table " + getTable());
 		CassandraCluster.getInstance().getSession()
-				.execute(getCreateStatement());
+				.execute(getRequeteCreation());
+		System.out.println("Table " + getTable() + " créée.");
 	}
 
 	protected abstract String getTable();
 
-	protected abstract String getCreateStatement();
+	protected abstract boolean doitEtreSupprimee();
+
+	protected abstract String getRequeteCreation();
 
 }
